Move LXD and LXDVM to checkbox-support
